feat(typescript): add typings to style utilities
